Split job_exec into multiple files
* Move message into its corresponding cmd type

* Locate generic message structures into core/message.go

* Implement helper functions for debug, info and error
