+++ core/formatter.go
@@ -1,4 +1,4 @@
-package message
+package core
 
 import (
 	"fmt"
@@ -15,7 +15,7 @@ var humanDivisors = [...]struct {
 	{"T", 1 << 40},
 }
 
-// HumanizeBytes takes a byte-size and returns a human-readable string
+// humanizeBytes takes a byte-size and returns a human-readable string
 func humanizeBytes(b int64) string {
 	var (
 		suffix string
