+++ core/message.go
@@ -1,0 +1,116 @@
+package core
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+
+	"github.com/peak/s5cmd/log"
+	"github.com/peak/s5cmd/objurl"
+	"github.com/peak/s5cmd/storage"
+)
+
+// InfoMessage is a generic message structure for successful operations.
+type InfoMessage struct {
+	Operation   string            `json:"operation"`
+	Success     bool              `json:"success"`
+	Source      *objurl.ObjectURL `json:"source"`
+	Destination *objurl.ObjectURL `json:"destination,omitempty"`
+	Object      *storage.Object   `json:"object,omitempty"`
+}
+
+// String is the string representation of InfoMessage.
+func (u InfoMessage) String() string {
+	return fmt.Sprintf("%v %v", u.Operation, u.Source)
+}
+
+// JSON is the JSON representation of InfoMessage.
+func (u InfoMessage) JSON() string {
+	u.Success = true
+	bytes, _ := json.Marshal(u)
+	return string(bytes)
+}
+
+// ErrorMessage is a generic message structure for unsuccessful operations.
+type ErrorMessage struct {
+	Operation string `json:"operation"`
+	Job       string `json:"job"`
+	Err       string `json:"error"`
+	Format    string `json:"-"`
+}
+
+// String is the string representation of ErrorMessage.
+func (e ErrorMessage) String() string {
+	err := cleanupSpaces(e.Err)
+	return fmt.Sprintf(e.Format, e.Job, err)
+}
+
+// JSON is the JSON representation of ErrorMessage.
+func (e ErrorMessage) JSON() string {
+	e.Err = cleanupSpaces(e.Err)
+	b, _ := json.Marshal(e)
+	return string(b)
+}
+
+// newError creates new ErrorMessage.
+func newError(job *Job, err error, format string) ErrorMessage {
+	errStr := ""
+	if err != nil {
+		errStr = err.Error()
+	}
+
+	msg := ErrorMessage{
+		Operation: job.operation.String(),
+		Job:       job.String(),
+		Err:       errStr,
+		Format:    format,
+	}
+	return msg
+}
+
+// warningMessage is the helper function to create warning messages.
+func warningMessage(job *Job, err error) {
+	format := "%q (%v)"
+	msg := newError(job, err, format)
+	log.Logger.Warning(msg)
+}
+
+// errorMessage is the helper function to create error messages.
+func errorMessage(job *Job, err error) {
+	format := "%q %v"
+	msg := newError(job, err, format)
+	log.Logger.Error(msg)
+}
+
+// DebugMessage is a generic message structure for debugging logs.
+type DebugMessage struct {
+	Content string `json:"content"`
+}
+
+// String is the string representation of DebugMessage.
+func (d DebugMessage) String() string {
+	return d.Content
+}
+
+// JSON is the JSON representation of DebugMessage.
+func (d DebugMessage) JSON() string {
+	b, _ := json.Marshal(d)
+	return string(b)
+}
+
+// debug is the helper function to create warning messages.
+func debug(format string, args ...interface{}) {
+	content := fmt.Sprintf(format, args...)
+	msg := DebugMessage{Content: content}
+	log.Logger.Debug(msg)
+}
+
+// cleanupError converts multiline error messages generated by aws-sdk-go into
+// a single line.
+func cleanupSpaces(s string) string {
+	s = strings.Replace(s, "\n", " ", -1)
+	s = strings.Replace(s, "\t", " ", -1)
+	s = strings.Replace(s, "  ", " ", -1)
+	s = strings.TrimSpace(s)
+	return s
+}
