+++ log/log.go
@@ -6,15 +6,21 @@ import (
 	"os"
 
 	"github.com/peak/s5cmd/flags"
-	"github.com/peak/s5cmd/message"
 )
 
 // stdoutCh is used to synchronize writes to standard output. Multi-line
 // logging is not possible if all workers print logs at the same time.
 var stdoutCh = make(chan string, 10000)
 
+// Logger is the global logger.
 var Logger *logger
 
+// Init inits global logger.
+func Init() {
+	Logger = New()
+}
+
+// logLevel is the level of Logger.
 type logLevel int
 
 const (
@@ -24,10 +30,7 @@ const (
 	levelError
 )
 
-func Init() {
-	Logger = New()
-}
-
+// String returns the string representation of logLevel.
 func (l logLevel) String() string {
 	switch l {
 	case levelInfo:
@@ -43,6 +46,8 @@ func (l logLevel) String() string {
 	}
 }
 
+// levelFromString returns logLevel for given string. It
+// return `levelInfo` as a default.
 func levelFromString(s string) logLevel {
 	switch s {
 	case "debug":
@@ -58,12 +63,20 @@ func levelFromString(s string) logLevel {
 	}
 }
 
+// Message is an interface to print structured logs.
+type Message interface {
+	fmt.Stringer
+	JSON() string
+}
+
+// logger is a structure for logging messages.
 type logger struct {
 	donech chan struct{}
 	impl   *log.Logger
 	level  logLevel
 }
 
+// New creates new logger.
 func New() *logger {
 	level := levelFromString(*flags.LogLevel)
 	logger := &logger{
@@ -75,7 +88,8 @@ func New() *logger {
 	return logger
 }
 
-func (l *logger) printf(level logLevel, message message.Message) {
+// printf prints message according to the given level and message.
+func (l *logger) printf(level logLevel, message Message) {
 	if level < l.level {
 		return
 	}
@@ -89,22 +103,27 @@ func (l *logger) printf(level logLevel, message message.Message) {
 	}
 }
 
-func (l *logger) Debug(msg message.Message) {
+// Debug prints message in debug mode.
+func (l *logger) Debug(msg Message) {
 	l.printf(levelDebug, msg)
 }
 
-func (l *logger) Info(msg message.Message) {
+// Info prints message in info mode.
+func (l *logger) Info(msg Message) {
 	l.printf(levelInfo, msg)
 }
 
-func (l *logger) Warning(msg message.Message) {
+// Warning prints message in warning mode.
+func (l *logger) Warning(msg Message) {
 	l.printf(levelWarning, msg)
 }
 
-func (l *logger) Error(msg message.Message) {
+// Error prints message in error mode.
+func (l *logger) Error(msg Message) {
 	l.printf(levelError, msg)
 }
 
+// stdout listens for stdoutCh and logs messages.
 func (l *logger) stdout() {
 	defer close(l.donech)
 
@@ -113,6 +132,7 @@ func (l *logger) stdout() {
 	}
 }
 
+// Close closes logger and its channel.
 func (l *logger) Close() {
 	close(stdoutCh)
 	<-l.donech
