--- message/message.go
@@ -1,181 +1,0 @@
-package message
-
-import (
-	"encoding/json"
-	"fmt"
-	"strings"
-
-	"github.com/peak/s5cmd/objurl"
-	"github.com/peak/s5cmd/storage"
-)
-
-const dateFormat = "2006/01/02 15:04:05"
-const listFormat = "%19s %1s %-6s %12s %s"
-
-type Message interface {
-	fmt.Stringer
-	JSON() string
-}
-
-type List struct {
-	Object        *storage.Object `json:"object"`
-	ShowEtag      bool            `json:"-"`
-	ShowHumanized bool            `json:"-"`
-}
-
-func (l List) humanize() string {
-	var size string
-	if l.ShowHumanized {
-		size = humanizeBytes(l.Object.Size)
-	} else {
-		size = fmt.Sprintf("%d", l.Object.Size)
-	}
-	return size
-}
-
-func (l List) String() string {
-	if l.Object.Type.IsDir() {
-		s := fmt.Sprintf(
-			listFormat,
-			"",
-			"",
-			"",
-			"DIR",
-			l.Object.URL.Relative(),
-		)
-		return s
-	}
-
-	var etag string
-	if l.ShowEtag {
-		etag = l.Object.Etag
-	}
-
-	s := fmt.Sprintf(
-		listFormat,
-		l.Object.ModTime.Format(dateFormat),
-		l.Object.StorageClass.ShortCode(),
-		etag,
-		l.humanize(),
-		l.Object.URL.Relative(),
-	)
-	return s
-}
-
-func (l List) JSON() string {
-	b, _ := json.Marshal(l.Object)
-	return string(b)
-}
-
-type Size struct {
-	Source       string `json:"source"`
-	StorageClass string `json:"storage_class,omitempty"`
-	Count        int64  `json:"count"`
-	Size         int64  `json:"size"`
-
-	ShowHumanized bool `json:"-"`
-}
-
-func (s Size) humanize() string {
-	var size string
-	if s.ShowHumanized {
-		size = humanizeBytes(s.Size)
-	} else {
-		size = fmt.Sprintf("%d", s.Size)
-	}
-	return size
-}
-
-func (s Size) String() string {
-	storageCls := ""
-	if s.StorageClass != "" {
-		storageCls = fmt.Sprintf(" [%s]", s.StorageClass)
-	}
-	return fmt.Sprintf(
-		"%s bytes in %d objects: %s%s",
-		s.humanize(),
-		s.Count,
-		s.Source,
-		storageCls,
-	)
-}
-
-func (s Size) JSON() string {
-	bytes, _ := json.Marshal(s)
-	return string(bytes)
-}
-
-type Info struct {
-	Operation   string            `json:"operation"`
-	Success     bool              `json:"success"`
-	Source      *objurl.ObjectURL `json:"source"`
-	Destination *objurl.ObjectURL `json:"destination,omitempty"`
-	Object      *storage.Object   `json:"object,omitempty"`
-}
-
-func (u Info) String() string {
-	return fmt.Sprintf("%v %v", u.Operation, u.Source)
-}
-
-func (u Info) JSON() string {
-	u.Success = true
-	bytes, _ := json.Marshal(u)
-	return string(bytes)
-}
-
-type Error struct {
-	Operation string `json:"operation"`
-	Job       string `json:"job"`
-	Err       string `json:"error"`
-}
-
-func (e Error) String() string {
-	err := cleanupSpaces(e.Err)
-	return fmt.Sprintf("%q : %v", e.Job, err)
-}
-
-func (e Error) JSON() string {
-	e.Err = cleanupSpaces(e.Err)
-	b, _ := json.Marshal(e)
-	return string(b)
-}
-
-type Warning struct {
-	Operation string `json:"operation"`
-	Job       string `json:"job"`
-	Err       string `json:"error"`
-}
-
-func (w Warning) String() string {
-	err := cleanupSpaces(w.Err)
-	return fmt.Sprintf("%q (%v)", w.Job, err)
-}
-
-func (w Warning) JSON() string {
-	w.Err = cleanupSpaces(w.Err)
-	b, _ := json.Marshal(w)
-	return string(b)
-}
-
-type Debug struct {
-	Content string `json:"content"`
-}
-
-func (d Debug) String() string {
-	return d.Content
-}
-
-func (d Debug) JSON() string {
-	b, _ := json.Marshal(d)
-	return string(b)
-}
-
-// cleanupError converts multiline error messages generated by aws-sdk-go into
-// a single line.
-func cleanupSpaces(s string) string {
-	s = strings.Replace(s, "\n", " ", -1)
-	s = strings.Replace(s, "\t", " ", -1)
-	s = strings.Replace(s, "  ", " ", -1)
-	s = strings.TrimSpace(s)
-	return s
-}
